subsets of raw images for AOI
For mosaicking the raw Sentinel-1 data, we first need to convert the data files to the (.tif) format and perform the mosaicking.
BUT,... If we convert all the tiles to the TIF format and then mosaic them all, it will require a high computing power and a lot of time, depending on the size of the Sentinel imagery being used.
Hence, to overcome this problem, I created a Well-Known-Text (WKT) for the AOI shapefile boundary, and subset out the only portions from the images that intersects with the WKT of the AOI. This saves the time, storage and computing services for processing. 

# How is it effective
There might be the possibility that some satellite images contains only a small portions of the area of interest, so subsetting the AOI portion from the whole image saves the processing for the unrequired portion of the image. This saves a lot of time when we are dealing with a large amount of images or some big data.